feat(db): port dialect-specific code + tests to bun:sqlite (sync)
Convert async db.transaction(...) sites to bun:sqlite's synchronous model
(all/run/get) and SQLite SQL idioms:
- idempotency.ts: sync SELECT-then-INSERT (single-writer; drop savepoint/23505)
- food-resolution, db/symptoms, auth, gdpr, signup, join, log: sync transactions
- guidance queries: strip int/text casts; EXTRACT(EPOCH) becomes integer-ms math
- raw-SQL date params use getTime() (bun:sqlite cannot bind a Date)
- gdpr/symptoms count casts stripped; healthz uses db.get()
- tests: idempotency rewritten for single-writer; texture uses pragma_table_info;
  log.create uses a trigger (SQLite has no ALTER TABLE ADD CONSTRAINT); signup
  transaction mocks made sync; remove obsolete Postgres drain test
- scripts/migrate-pg-to-sqlite.ts: Postgres-JSON-snapshot to migrated SQLite ETL

Full suite green (180 files, 1350 pass, 0 fail); typecheck/lint/build clean.

+++ src/lib/server/guidance/queries/seasonal.ts
@@ -47,7 +47,7 @@ export async function loadSeasonalFoods(ageMonths: number, month: number): Promi
 
 export async function loadTexturesTried(childId: number): Promise<number> {
   const rows = await db
-    .select({ n: sql<number>`count(distinct ${foodEntries.texture})::int` })
+    .select({ n: sql<number>`count(distinct ${foodEntries.texture})` })
     .from(foodEntries)
     .where(and(eq(foodEntries.childId, childId), sql`${foodEntries.texture} IS NOT NULL`))
     .limit(1);
